+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/storage/StorageContainerItemHelper.java
@@ -4,14 +4,17 @@ import com.refinedmods.refinedstorage2.api.storage.Storage;
 import com.refinedmods.refinedstorage2.api.storage.StorageInfo;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
-import java.util.UUID;
+import java.util.Set;
 import java.util.function.LongFunction;
 import javax.annotation.Nullable;
 
 import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
@@ -38,9 +41,13 @@ public interface StorageContainerItemHelper {
                          LongFunction<String> amountFormatter,
                          boolean hasCapacity);
 
-    // TODO: remove - leaky abstraction
-    Optional<UUID> getId(ItemStack stack);
+    void transferToBlockEntity(ItemStack stack, ItemTransferableStorageBlockEntity blockEntity);
 
-    // TODO: remove - leaky abstraction
-    void setId(ItemStack stack, UUID id);
+    void transferFromBlockEntity(ItemStack stack, ItemTransferableStorageBlockEntity blockEntity);
+
+    void registerDiskModel(Item item, ResourceLocation model);
+
+    Set<ResourceLocation> getDiskModels();
+
+    Map<Item, ResourceLocation> getDiskModelsByItem();
 }