+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/storage/storageblock/FluidStorageBlockBlockItem.java
@@ -7,20 +7,13 @@ import com.refinedmods.refinedstorage2.platform.common.content.Items;
 import com.refinedmods.refinedstorage2.platform.common.storage.FluidStorageType;
 import com.refinedmods.refinedstorage2.platform.common.support.resource.FluidResourceRendering;
 
-import java.util.UUID;
 import javax.annotation.Nullable;
 
-import net.minecraft.core.BlockPos;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.entity.BlockEntity;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class FluidStorageBlockBlockItem extends AbstractStorageContainerBlockItem {
-    private static final Logger LOGGER = LoggerFactory.getLogger(FluidStorageBlockBlockItem.class);
-
     private final FluidStorageType.Variant variant;
 
     public FluidStorageBlockBlockItem(final Block block, final FluidStorageType.Variant variant) {
@@ -54,17 +47,5 @@ public class FluidStorageBlockBlockItem extends AbstractStorageContainerBlockIte
             return null;
         }
         return new ItemStack(Items.INSTANCE.getFluidStoragePart(variant), count);
-    }
-
-    @Override
-    protected void updateBlockEntityWithStorageId(final BlockPos pos,
-                                                  @Nullable final BlockEntity blockEntity,
-                                                  final UUID id) {
-        if (blockEntity instanceof AbstractStorageBlockBlockEntity<?> storageBlockEntity) {
-            LOGGER.debug("Transferred storage {} to block at {}", id, pos);
-            storageBlockEntity.modifyStorageIdAfterAlreadyInitialized(id);
-        } else {
-            LOGGER.warn("Storage {} could not be set, block entity does not exist yet at {}", id, pos);
-        }
     }
 }