+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/storage/diskdrive/DiskDriveBakedModel.java
@@ -1,9 +1,9 @@
 package com.refinedmods.refinedstorage2.platform.fabric.storage.diskdrive;
 
-import com.refinedmods.refinedstorage2.api.network.impl.node.multistorage.MultiStorageState;
 import com.refinedmods.refinedstorage2.api.network.impl.node.multistorage.MultiStorageStorageState;
 import com.refinedmods.refinedstorage2.platform.common.storage.diskdrive.AbstractDiskDriveBlockEntity;
 import com.refinedmods.refinedstorage2.platform.common.storage.diskdrive.DiskDriveBlock;
+import com.refinedmods.refinedstorage2.platform.common.storage.diskdrive.DiskDriveDisk;
 import com.refinedmods.refinedstorage2.platform.common.support.direction.BiDirection;
 import com.refinedmods.refinedstorage2.platform.fabric.support.render.QuadRotator;
 import com.refinedmods.refinedstorage2.platform.fabric.support.render.QuadTranslator;
@@ -11,6 +11,7 @@ import com.refinedmods.refinedstorage2.platform.fabric.support.render.QuadTransl
 import java.util.EnumMap;
 import java.util.Map;
 import java.util.function.Supplier;
+import javax.annotation.Nullable;
 
 import net.fabricmc.fabric.api.renderer.v1.model.ForwardingBakedModel;
 import net.fabricmc.fabric.api.renderer.v1.render.RenderContext;
@@ -20,6 +21,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.BlockAndTintGetter;
 import net.minecraft.world.level.block.state.BlockState;
@@ -45,15 +47,15 @@ class DiskDriveBakedModel extends ForwardingBakedModel {
         }
     }
 
-    private final BakedModel diskModel;
-    private final BakedModel diskInactiveModel;
+    private final Map<Item, BakedModel> diskModels;
+    private final BakedModel inactiveLedModel;
 
     DiskDriveBakedModel(final BakedModel baseModel,
-                        final BakedModel diskModel,
-                        final BakedModel diskInactiveModel) {
+                        final Map<Item, BakedModel> diskModels,
+                        final BakedModel inactiveLedModel) {
         this.wrapped = baseModel;
-        this.diskModel = diskModel;
-        this.diskInactiveModel = diskInactiveModel;
+        this.diskModels = diskModels;
+        this.inactiveLedModel = inactiveLedModel;
     }
 
     @Override
@@ -71,12 +73,8 @@ class DiskDriveBakedModel extends ForwardingBakedModel {
             return;
         }
         for (int i = 0; i < TRANSLATORS.length; ++i) {
-            if (!AbstractDiskDriveBlockEntity.hasDisk(tag, i)) {
-                continue;
-            }
-            context.pushTransform(TRANSLATORS[i]);
-            diskInactiveModel.emitItemQuads(stack, randomSupplier, context);
-            context.popTransform();
+            final Item diskItem = AbstractDiskDriveBlockEntity.getDisk(tag, i);
+            emitDiskQuads(stack, randomSupplier, context, diskItem, i);
         }
     }
 
@@ -101,8 +99,8 @@ class DiskDriveBakedModel extends ForwardingBakedModel {
 
         if (blockView instanceof RenderAttachedBlockView renderAttachedBlockView) {
             final Object renderAttachment = renderAttachedBlockView.getBlockEntityRenderAttachment(pos);
-            if (renderAttachment instanceof MultiStorageState states) {
-                emitDiskQuads(blockView, state, pos, randomSupplier, context, states);
+            if (renderAttachment instanceof DiskDriveDisk[] disks) {
+                emitDiskQuads(blockView, state, pos, randomSupplier, context, disks);
             }
         }
 
@@ -114,14 +112,48 @@ class DiskDriveBakedModel extends ForwardingBakedModel {
                                final BlockPos pos,
                                final Supplier<RandomSource> randomSupplier,
                                final RenderContext context,
-                               final MultiStorageState states) {
+                               final DiskDriveDisk[] disks) {
         for (int i = 0; i < TRANSLATORS.length; ++i) {
-            if (states.getState(i) == MultiStorageStorageState.NONE) {
-                continue;
-            }
-            context.pushTransform(TRANSLATORS[i]);
-            diskModel.emitBlockQuads(blockView, state, pos, randomSupplier, context);
-            context.popTransform();
+            final DiskDriveDisk disk = disks[i];
+            emitDiskQuads(blockView, state, pos, randomSupplier, context, disk, i);
         }
+    }
+
+    private void emitDiskQuads(final BlockAndTintGetter blockView,
+                               final BlockState state,
+                               final BlockPos pos,
+                               final Supplier<RandomSource> randomSupplier,
+                               final RenderContext context,
+                               final DiskDriveDisk disk,
+                               final int index) {
+        if (disk.state() == MultiStorageStorageState.NONE) {
+            return;
+        }
+        final BakedModel model = diskModels.get(disk.item());
+        if (model == null) {
+            return;
+        }
+        context.pushTransform(TRANSLATORS[index]);
+        model.emitBlockQuads(blockView, state, pos, randomSupplier, context);
+        context.popTransform();
+    }
+
+
+    private void emitDiskQuads(final ItemStack stack,
+                               final Supplier<RandomSource> randomSupplier,
+                               final RenderContext context,
+                               @Nullable final Item diskItem,
+                               final int index) {
+        if (diskItem == null) {
+            return;
+        }
+        final BakedModel diskModel = diskModels.get(diskItem);
+        if (diskModel == null) {
+            return;
+        }
+        context.pushTransform(TRANSLATORS[index]);
+        diskModel.emitItemQuads(stack, randomSupplier, context);
+        inactiveLedModel.emitItemQuads(stack, randomSupplier, context);
+        context.popTransform();
     }
 }