+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/storage/diskdrive/DiskDriveUnbakedGeometry.java
@@ -1,10 +1,13 @@
 package com.refinedmods.refinedstorage2.platform.forge.storage.diskdrive;
 
+import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
 import com.refinedmods.refinedstorage2.platform.common.support.direction.BiDirection;
 
+import java.util.Map;
 import java.util.Objects;
 import java.util.function.BiFunction;
 import java.util.function.Function;
+import java.util.stream.Collectors;
 
 import com.mojang.math.Transformation;
 import net.minecraft.client.renderer.block.model.ItemOverrides;
@@ -23,9 +26,8 @@ import org.joml.Vector3f;
 import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createIdentifier;
 
 public class DiskDriveUnbakedGeometry implements IUnbakedGeometry<DiskDriveUnbakedGeometry> {
-    private static final ResourceLocation BASE_MODEL = createIdentifier("block/disk_drive_base");
-    private static final ResourceLocation DISK_MODEL = createIdentifier("block/disk");
-    private static final ResourceLocation DISK_INACTIVE_MODEL = createIdentifier("block/disk_inactive");
+    private static final ResourceLocation BASE_MODEL = createIdentifier("block/disk_drive/base");
+    private static final ResourceLocation LED_INACTIVE_MODEL = createIdentifier("block/disk/led_inactive");
 
     DiskDriveUnbakedGeometry() {
     }
@@ -34,8 +36,10 @@ public class DiskDriveUnbakedGeometry implements IUnbakedGeometry<DiskDriveUnbak
     public void resolveParents(final Function<ResourceLocation, UnbakedModel> modelGetter,
                                final IGeometryBakingContext context) {
         modelGetter.apply(BASE_MODEL).resolveParents(modelGetter);
-        modelGetter.apply(DISK_MODEL).resolveParents(modelGetter);
-        modelGetter.apply(DISK_INACTIVE_MODEL).resolveParents(modelGetter);
+        PlatformApi.INSTANCE.getStorageContainerItemHelper().getDiskModels().forEach(
+            diskModel -> modelGetter.apply(diskModel).resolveParents(modelGetter)
+        );
+        modelGetter.apply(LED_INACTIVE_MODEL).resolveParents(modelGetter);
     }
 
     @Override
@@ -45,11 +49,28 @@ public class DiskDriveUnbakedGeometry implements IUnbakedGeometry<DiskDriveUnbak
                            final ModelState modelState,
                            final ItemOverrides overrides,
                            final ResourceLocation modelLocation) {
+        final var diskModelBakeries = PlatformApi.INSTANCE.getStorageContainerItemHelper()
+            .getDiskModelsByItem()
+            .entrySet()
+            .stream()
+            .collect(Collectors.toMap(
+                Map.Entry::getKey,
+                entry -> getDiskModelBakery(modelState, baker, spriteGetter, entry.getValue())
+            ));
+        final var diskItemModelBakeries = PlatformApi.INSTANCE.getStorageContainerItemHelper()
+            .getDiskModelsByItem()
+            .entrySet()
+            .stream()
+            .collect(Collectors.toMap(
+                Map.Entry::getKey,
+                entry -> getItemModelBakery(modelState, baker, spriteGetter, entry.getValue())
+            ));
         return new DiskDriveBakedModel(
             getBaseModelBakery(modelState, baker, spriteGetter),
             Objects.requireNonNull(baker.bake(BASE_MODEL, modelState, spriteGetter)),
-            getDiskModelBakery(modelState, baker, spriteGetter),
-            getDiskItemModelBakery(modelState, baker, spriteGetter)
+            diskModelBakeries,
+            diskItemModelBakeries,
+            getItemModelBakery(modelState, baker, spriteGetter, LED_INACTIVE_MODEL)
         );
     }
 
@@ -66,23 +87,24 @@ public class DiskDriveUnbakedGeometry implements IUnbakedGeometry<DiskDriveUnbak
     private BiFunction<BiDirection, Vector3f, BakedModel> getDiskModelBakery(final ModelState state,
                                                                              final ModelBaker baker,
                                                                              final Function
-                                                                                 <Material, TextureAtlasSprite> sg) {
+                                                                                 <Material, TextureAtlasSprite> sg,
+                                                                             final ResourceLocation model) {
         return (direction, trans) -> {
             final Transformation translation = new Transformation(trans, null, null, null);
             final Transformation rotation = new Transformation(null, direction.getQuaternion(), null, null);
             final ModelState wrappedState = new SimpleModelState(rotation.compose(translation), state.isUvLocked());
-            return baker.bake(DISK_MODEL, wrappedState, sg);
+            return baker.bake(model, wrappedState, sg);
         };
     }
 
-    private Function<Vector3f, BakedModel> getDiskItemModelBakery(final ModelState state,
-                                                                  final ModelBaker baker,
-                                                                  final Function<Material, TextureAtlasSprite>
-                                                                      sg) {
+    private Function<Vector3f, BakedModel> getItemModelBakery(final ModelState state,
+                                                              final ModelBaker baker,
+                                                              final Function<Material, TextureAtlasSprite> sg,
+                                                              final ResourceLocation model) {
         return trans -> {
             final Transformation translation = new Transformation(trans, null, null, null);
             final ModelState wrappedState = new SimpleModelState(translation, state.isUvLocked());
-            return baker.bake(DISK_INACTIVE_MODEL, wrappedState, sg);
+            return baker.bake(model, wrappedState, sg);
         };
     }
 }