+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+-   Custom disk models. Fluid disks now have a different model.
+
 ## [2.0.0-milestone.3.2] - 2023-11-03
 
 ### Added